backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==1.0.0
openai==1.8.0
gunicorn==20.1.0
werkzeug==2.0.2 